Extract leaflet JS from geoblacklight JS
