--- app/javascript/geoblacklight/controls/opacity.js
@@ -1,6 +1,0 @@
-import GeoBlacklightLeafletPlugin from '../leaflet_plugin.js'
-
-// Leaflet opacity control.
-GeoBlacklightLeafletPlugin.Controls.Opacity = function() {
-  this.map.addControl(new L.Control.LayerOpacity(this.overlay));
-};
