+++ app/javascript/geoblacklight/geoblacklight_initializer.js
@@ -1,0 +1,13 @@
+import initializePopovers from "./popovers.js";
+import initializeRelations from "./relations.js";
+import initializeTooltips from "./tooltips.js";
+import initializeTruncation from "./truncation.js";
+
+export default class GeoBlacklightInitializer {
+  run() {
+    initializeRelations();
+    initializeTruncation();
+    initializePopovers();
+    initializeTooltips();
+  }
+}
