+++ app/javascript/geoblacklight/metadata.js
