+++ app/javascript/leaflet/downloaders/downloader.js
