+++ app/javascript/leaflet/downloaders/hgl_downloader.js
