+++ app/javascript/leaflet/viewers/cog.js
