+++ app/javascript/leaflet/viewers/esri/dynamic_map_layer.js
