+++ app/javascript/leaflet/viewers/esri/image_map_layer.js
