+++ app/javascript/leaflet/viewers/esri/tiled_map_layer.js
