+++ app/javascript/leaflet/viewers/iiif.js
