+++ app/javascript/leaflet/viewers/index_map.js
