+++ app/javascript/leaflet/viewers/tilejson.js
