+++ app/javascript/leaflet/viewers/xyz.js
