+++ app/javascript/stylesheets/leaflet.css
@@ -1,0 +1,2 @@
+@import "../../../node_modules/leaflet/dist/leaflet.css";
+@import "../../../node_modules/leaflet-fullscreen/dist/leaflet.fullscreen.css";
