+++ lib/generators/geoblacklight/templates/geoblacklight.js
@@ -1,6 +1,6 @@
-import '@geoblacklight/frontend/dist/style.css'
-import { GeoBlacklightLeafletPlugin } from '@geoblacklight/frontend'
+import "@geoblacklight/frontend/dist/style.css";
+import { GeoBlacklightInitializer } from "@geoblacklight/frontend";
 
-document.addEventListener('DOMContentLoaded', () => {
-  console.log('loaded')
-})
+document.addEventListener("DOMContentLoaded", () => {
+  new GeoBlacklightInitializer().run();
+});
