+++ lib/generators/geoblacklight/templates/leaflet.js
@@ -1,0 +1,5 @@
+import { LeafletInitializer } from "@geoblacklight/frontend";
+
+document.addEventListener("DOMContentLoaded", () => {
+  new LeafletInitializer().run();
+});
